Изменения по замечаниям в кодревью
