+++ core/build.gradle.kts
@@ -1,4 +1,3 @@
-
 plugins {
     id(Config.Plugins.androidLibrary)
     id(Config.Plugins.androidKotlin)
