+++ core/src/main/java/com/sinx/core/Feature.kt
@@ -1,5 +1,6 @@
+package com.sinx.core
 import javax.inject.Scope
 
 @Scope
 @Retention(AnnotationRetention.RUNTIME)
-annotation class AppScope
+annotation class Feature
