+++ task/src/main/java/com/sinx/task/di/TaskDeps.kt
@@ -1,0 +1,7 @@
+package com.sinx.task.di
+
+import com.sinx.coredbinterface.dao.TaskDAO
+
+interface TaskDeps {
+    val taskDAO: TaskDAO
+}
