+++ task/src/main/java/com/sinx/task/di/TaskDepsStore.kt
@@ -1,0 +1,7 @@
+package com.sinx.task.di
+
+import kotlin.properties.Delegates
+
+object TaskDepsStore : TaskDepsProvider {
+    override var deps: TaskDeps by Delegates.notNull()
+}
